backend/requirements.txt
flask-cors==6.0.1
PyJWT==2.8.0
flask==3.0.0
